+++ whatsapp_addon/index.js
@@ -2,8 +2,8 @@ const express = require("express");
 const bodyParser = require("body-parser");
 const cors = require("cors");
 const axios = require("axios");
-const fs = require('fs');
-const WhatsappClient = require('./WhatsappClient.js')
+const fs = require("fs");
+const WhatsappClient = require("./WhatsappClient.js");
 
 var logger = require("log4js").getLogger();
 logger.level = "info";
@@ -17,21 +17,23 @@ app.use(cors());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 
-fs.readFile('data/options.json', function (error, content) {
+fs.readFile("data/options.json", function (error, content) {
   var options = JSON.parse(content);
   const clients = {};
 
   options.clients.forEach((client) => {
-    clients[client] = new WhatsappClient(`/data/${client}.json`)
-  })
+    clients[client] = new WhatsappClient(`/data/${client}.json`);
+  });
 
   // Load clients
   Object.keys(clients).forEach(function (key, index) {
-    const wapp = clients[key]
+    const wapp = clients[key];
 
     wapp.on("qr", (qr) => {
-      logger.info(key, "require authentication over QRCode, please see your notifications...");
-
+      logger.info(
+        key,
+        "require authentication over QRCode, please see your notifications..."
+      );
 
       var code = qrimage.image(qr, { type: "png" });
 
@@ -51,11 +53,11 @@ fs.readFile('data/options.json', function (error, content) {
           }
         );
       });
-
     });
 
     wapp.on("ready", () => {
-      logger.info(key, "client is ready.")
+      logger.info(key, "client is ready.");
+      wapp.sendPresenceUpdate("unavailable");
       axios.post(
         "http://supervisor/core/api/services/persistent_notification/dismiss",
         {
@@ -70,18 +72,35 @@ fs.readFile('data/options.json', function (error, content) {
     });
 
     wapp.on("message", (msg) => {
-      axios.post("http://supervisor/core/api/events/new_whatsapp_message", { clientId: key, ...msg }, {
-        headers: {
-          Authorization: `Bearer ${process.env.SUPERVISOR_TOKEN}`,
-        },
-      });
+      axios.post(
+        "http://supervisor/core/api/events/new_whatsapp_message",
+        { clientId: key, ...msg },
+        {
+          headers: {
+            Authorization: `Bearer ${process.env.SUPERVISOR_TOKEN}`,
+          },
+        }
+      );
       logger.debug(`New message event fired from ${key}.`);
     });
 
-    wapp.on('disconnected', () => {
+    wapp.on("disconnected", () => {
       logger.info(`Client ${key} was logged out. Restart to add it again.`);
       delete clients[key];
       fs.unlinkSync(`/data/${key}.json`);
+    });
+
+    wapp.on("presence_update", (presence) => {
+      axios.post(
+        "http://supervisor/core/api/events/whatsapp_presence_update",
+        { clientId: key, ...presence },
+        {
+          headers: {
+            Authorization: `Bearer ${process.env.SUPERVISOR_TOKEN}`,
+          },
+        }
+      );
+      logger.debug(`New presence event fired from ${key}.`);
     });
   });
 
@@ -89,41 +108,77 @@ fs.readFile('data/options.json', function (error, content) {
 
   app.post("/sendMessage", (req, res) => {
     const message = req.body;
-    if (message.hasOwnProperty('clientId')) {
+    if (message.hasOwnProperty("clientId")) {
       if (clients.hasOwnProperty(message.clientId)) {
         const wapp = clients[message.clientId];
-        wapp.sendMessage(message.to, message.body, message.options).then(() => {
-          res.send("OK");
-          logger.debug("Message successfully sended from addon.");
-        }).catch((error) => {
-          res.send("KO");
-          logger.error(error);
-        });
+        wapp
+          .sendMessage(message.to, message.body, message.options)
+          .then(() => {
+            res.send("OK");
+            logger.debug("Message successfully sended from addon.");
+          })
+          .catch((error) => {
+            res.send("KO");
+            logger.error(error);
+          });
       } else {
-        logger.error('Error in sending message. Client ID not found.')
+        logger.error("Error in sending message. Client ID not found.");
       }
     } else {
-      logger.error('Error in sending message. Please specify client ID.')
+      logger.error("Error in sending message. Please specify client ID.");
     }
   });
 
   app.post("/setStatus", (req, res) => {
-    const status = req.body.status
-    if (req.body.hasOwnProperty('clientId')) {
+    const status = req.body.status;
+    if (req.body.hasOwnProperty("clientId")) {
       if (clients.hasOwnProperty(req.body.clientId)) {
         const wapp = clients[req.body.clientId];
 
-        wapp.up
-
         wapp.updateProfileStatus(status).then(() => {
-          res.send('OK')
-        })
+          res.send("OK");
+        });
       } else {
-        logger.error('Error in set status. Client ID not found.')
+        logger.error("Error in set status. Client ID not found.");
       }
     } else {
-      logger.error('Error in set status. Please specify client ID.')
+      logger.error("Error in set status. Please specify client ID.");
     }
+  });
 
-  })
+  app.post("/presenceSubscribe", (req, res) => {
+    const request = req.body;
+
+    if (req.body.hasOwnProperty("clientId")) {
+      if (clients.hasOwnProperty(req.body.clientId)) {
+        const wapp = clients[req.body.clientId];
+
+        wapp.presenceSubscribe(request.userId).then(() => {
+          res.send("OK");
+        });
+      } else {
+        logger.error("Error in subscribe presence. Client ID not found.");
+      }
+    } else {
+      logger.error("Error in subscribe presence. Please specify client ID.");
+    }
+  });
+
+  app.post("/sendPresenceUpdate", (req, res) => {
+    const request = req.body;
+
+    if (req.body.hasOwnProperty("clientId")) {
+      if (clients.hasOwnProperty(req.body.clientId)) {
+        const wapp = clients[req.body.clientId];
+
+        wapp.sendPresenceUpdate(request.type, request.to).then(() => {
+          res.send("OK");
+        });
+      } else {
+        logger.error("Error in presence update. Client ID not found.");
+      }
+    } else {
+      logger.error("Error in presence update. Please specify client ID.");
+    }
+  });
 });